Add files via upload
This is the initial version of the function which sorts the tickers into three categories of beats, meets, and misses.
The code is tested for a small .csv file (first 30 of the R3000 file). And it gives the correct output for that small file.
Changes would be made to the inputData() function to work on the R3000 file.